Create ProProfs AWS ML Speciality
Hello,

I benefited a lot from this repo in my AWS ML Specialty preparations, and I thought why not add to it if i have something to add?
Here's a link to a google document in which i gathered ProProfs 180 Questions with their ProProfs official answers (which could be wrong) and some random comments.

I can enable the editing in the document if someone wants to add some comments or notes. Just contact me: yomnaj96@gmail.com

Thank you for helping me pass the exam, and wish you the best of luck.